feat: add home page with Clerk SignIn component and logo refactor: update dashboard page to include ImageProcessor component refactor: modify middleware to protect routes except public Clerk paths
