+++ apps/frontend/middleware.ts
@@ -4,9 +4,9 @@ export default clerkMiddleware();
 
 export const config = {
   matcher: [
-    // Skip Next.js internals and all static files, unless found in search params
-    "/((?!_next|[^?]*\\.(?:html?|css|js(?!on)|jpe?g|webp|png|gif|svg|ttf|woff2?|ico|csv|docx?|xlsx?|zip|webmanifest)).*)",
-    // Always run for API routes
-    "/(api|trpc)(.*)",
+    // Proteger todo excepto rutas públicas de Clerk
+    "/((?!_next|favicon.ico|sign-up|forgot-password|api|\\[\\[...rest\\]\\]|/$).*)",
+    // Siempre proteger rutas de API si las tienes aquí
+    // "/(api|trpc)(.*)",
   ],
 };
